dropIndex() kon een fatal error geven als de tabel waar de index op had moeten zitten ook al niet bestond, fixes 802

+++ lib/dbstruct/SpotStruct_mysql.php
@@ -65,6 +65,12 @@ class SpotStruct_mysql extends SpotStruct_abs {
 
 	/* dropt een index als deze bestaat */
 	function dropIndex($idxname, $tablename) {
+		# Check eerst of de tabel bestaat, anders kan
+		# indexExists mislukken en een fatal error geven
+		if (!$this->tableExists($tablename)) {
+			return ;
+		} # if
+		
 		if ($this->indexExists($idxname, $tablename)) {
 			$this->_dbcon->rawExec("DROP INDEX " . $idxname . " ON " . $tablename);
 		} # if
